+++ account-keeper-impl/src/main/java/com/dwarfeng/acckeeper/impl/service/telqos/ProtectLocalCacheCommand.java
@@ -1,0 +1,109 @@
+package com.dwarfeng.acckeeper.impl.service.telqos;
+
+import com.dwarfeng.acckeeper.stack.handler.Protector;
+import com.dwarfeng.acckeeper.stack.service.ProtectQosService;
+import com.dwarfeng.springtelqos.node.config.TelqosCommand;
+import com.dwarfeng.springtelqos.sdk.command.CliCommand;
+import com.dwarfeng.springtelqos.stack.command.Context;
+import com.dwarfeng.springtelqos.stack.exception.TelqosException;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import org.apache.commons.cli.CommandLine;
+import org.apache.commons.cli.Option;
+import org.apache.commons.cli.ParseException;
+import org.apache.commons.lang3.tuple.Pair;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+
+@TelqosCommand
+public class ProtectLocalCacheCommand extends CliCommand {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(ProtectLocalCacheCommand.class);
+
+    private static final String COMMAND_OPTION_LOOKUP = "l";
+    private static final String COMMAND_OPTION_CLEAR = "c";
+
+    private static final String[] COMMAND_OPTION_ARRAY = new String[]{
+            COMMAND_OPTION_LOOKUP,
+            COMMAND_OPTION_CLEAR
+    };
+
+    private static final String IDENTITY = "plc";
+    private static final String DESCRIPTION = "数据记录本地缓存操作";
+    private static final String CMD_LINE_SYNTAX_LOOKUP = IDENTITY + " " +
+            CommandUtil.concatOptionPrefix(COMMAND_OPTION_LOOKUP) + " account-id";
+    private static final String CMD_LINE_SYNTAX_CLEAR = IDENTITY + " " +
+            CommandUtil.concatOptionPrefix(COMMAND_OPTION_CLEAR);
+
+    private static final String[] CMD_LINE_ARRAY = new String[]{
+            CMD_LINE_SYNTAX_LOOKUP,
+            CMD_LINE_SYNTAX_CLEAR
+    };
+
+    private static final String CMD_LINE_SYNTAX = CommandUtil.syntax(CMD_LINE_ARRAY);
+
+    private final ProtectQosService protectQosService;
+
+    public ProtectLocalCacheCommand(ProtectQosService protectQosService) {
+        super(IDENTITY, DESCRIPTION, CMD_LINE_SYNTAX);
+        this.protectQosService = protectQosService;
+    }
+
+    @Override
+    protected List<Option> buildOptions() {
+        List<Option> list = new ArrayList<>();
+        list.add(Option.builder(COMMAND_OPTION_LOOKUP).optionalArg(true).hasArg(true).type(String.class)
+                .argName("account-id").desc("查看指定数据点的详细信息，如果本地缓存中不存在，则尝试抓取").build());
+        list.add(Option.builder(COMMAND_OPTION_CLEAR).optionalArg(true).hasArg(false).desc("清除缓存").build());
+        return list;
+    }
+
+    @SuppressWarnings("DuplicatedCode")
+    @Override
+    protected void executeWithCmd(Context context, CommandLine cmd) throws TelqosException {
+        try {
+            Pair<String, Integer> pair = CommandUtil.analyseCommand(cmd, COMMAND_OPTION_ARRAY);
+            if (pair.getRight() != 1) {
+                context.sendMessage(CommandUtil.optionMismatchMessage(COMMAND_OPTION_ARRAY));
+                context.sendMessage(CMD_LINE_SYNTAX);
+                return;
+            }
+            switch (pair.getLeft()) {
+                case COMMAND_OPTION_LOOKUP:
+                    handleLookup(context, cmd);
+                    break;
+                case COMMAND_OPTION_CLEAR:
+                    handleClear(context);
+                    break;
+            }
+        } catch (Exception e) {
+            throw new TelqosException(e);
+        }
+    }
+
+    private void handleLookup(Context context, CommandLine cmd) throws Exception {
+        String accountId;
+        try {
+            accountId = ((String) cmd.getParsedOptionValue(COMMAND_OPTION_LOOKUP));
+        } catch (ParseException e) {
+            LOGGER.warn("解析命令选项时发生异常，异常信息如下", e);
+            context.sendMessage("命令行格式错误，正确的格式为: " + CMD_LINE_SYNTAX_LOOKUP);
+            context.sendMessage("请留意选项 p 后接参数的类型应该是数字 ");
+            return;
+        }
+        Protector protector = protectQosService.getProtector(new StringIdKey(accountId));
+        if (Objects.isNull(protector)) {
+            context.sendMessage("not exists!");
+            return;
+        }
+        context.sendMessage(Objects.toString(protector));
+    }
+
+    private void handleClear(Context context) throws Exception {
+        protectQosService.clearLocalCache();
+        context.sendMessage("缓存已清空");
+    }
+}